rename token gate
