+++ components/common/SpaceAccessGate/components/TokenGate/TokenGateContent.tsx
