+++ components/common/SpaceAccessGate/components/TokenGate/TokenGateOption.tsx
