+++ components/common/SpaceAccessGate/components/TokenGate/VerifyCheckmark.tsx
